+++ rut/src/main/java/io/norberg/rut/Router.java
@@ -170,14 +170,14 @@ public final class Router<T> {
    */
   public static class Result<T> {
 
-    private final Captor captor;
+    private final Trie.Captor captor;
 
     private Status status;
     private Target<T> target;
     private CharSequence path;
 
     private Result(final int captures) {
-      captor = new Captor(captures);
+      captor = new Trie.Captor(captures);
     }
 
     /**
