+++ rut/src/main/java/io/norberg/rut/Trie.java
@@ -170,4 +170,70 @@ class Trie<T> {
     }
     return -1;
   }
+
+  static class Captor {
+
+    private final int[] start;
+    private final int[] end;
+    private boolean match;
+    private int captured;
+
+    Captor(final int captures) {
+      this.start = new int[captures];
+      this.end = new int[captures];
+    }
+
+    void reset() {
+      match = false;
+      captured = 0;
+    }
+
+    void capture(final int i, final int start, final int end) {
+      this.start[i] = start;
+      this.end[i] = end;
+    }
+
+    void match(final int captured) {
+      match = true;
+      this.captured = captured;
+    }
+
+    boolean isMatch() {
+      return match;
+    }
+
+    int values() {
+      return captured;
+    }
+
+    int valueStart(final int i) {
+      if (!match) {
+        throw new IllegalStateException("not matched");
+      }
+      if (i > captured) {
+        throw new IndexOutOfBoundsException();
+      }
+      return start[i];
+    }
+
+    int valueEnd(final int i) {
+      if (!match) {
+        throw new IllegalStateException("not matched");
+      }
+      if (i > captured) {
+        throw new IndexOutOfBoundsException();
+      }
+      return end[i];
+    }
+
+    CharSequence value(final CharSequence haystack, final int i) {
+      if (!match) {
+        throw new IllegalStateException("not matched");
+      }
+      if (i > captured) {
+        throw new IndexOutOfBoundsException();
+      }
+      return haystack.subSequence(start[i], end[i]);
+    }
+  }
 }
